update to nw v4.1 (blender 5.1)
