+++ node_wrangler_octane/operators/change_factor.py
@@ -1,0 +1,51 @@
+# SPDX-FileCopyrightText: 2025 Blender Foundation
+#
+# SPDX-License-Identifier: GPL-2.0-or-later
+
+import bpy
+from bpy.types import Operator
+from bpy.props import FloatProperty
+
+from ..utils.nodes import (
+    NWBase,
+    nw_check,
+    nw_check_selected,
+    get_nodes_links,
+)
+
+
+#### ------------------------------ OPERATORS ------------------------------ ####
+
+class NODE_OT_change_factor(Operator, NWBase):
+    bl_idname = "node.nw_factor"
+    bl_label = "Change Factor"
+    bl_description = "Change factors of Mix nodes and Mix Shader nodes"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    @classmethod
+    def poll(cls, context):
+        return nw_check(cls, context) and nw_check_selected(cls, context)
+
+    # option: Change factor.
+    # If option is 1.0 or 0.0 - set to 1.0 or 0.0
+    # Else - change factor by option value.
+    option: FloatProperty()
+
+    def execute(self, context):
+        nodes, links = get_nodes_links(context)
+        option = self.option
+        selected = []  # entry = index
+        for si, node in enumerate(nodes):
+            if node.select:
+                if node.type in {'MIX_RGB', 'MIX_SHADER'} or node.bl_idname == 'ShaderNodeMix':
+                    selected.append(si)
+
+        for si in selected:
+            fac = nodes[si].inputs[0]
+            nodes[si].hide = False
+            if option in {0.0, 1.0}:
+                fac.default_value = option
+            else:
+                fac.default_value += option
+
+        return {'FINISHED'}
