+++ node_wrangler_octane/operators/labels_clear.py
@@ -1,0 +1,42 @@
+# SPDX-FileCopyrightText: 2025 Blender Foundation
+#
+# SPDX-License-Identifier: GPL-2.0-or-later
+
+import bpy
+from bpy.types import Operator
+from bpy.props import BoolProperty
+
+from ..utils.nodes import (
+    NWBase,
+    nw_check,
+    nw_check_selected,
+    get_nodes_links,
+)
+
+
+#### ------------------------------ OPERATORS ------------------------------ ####
+
+class NODE_OT_labels_clear(Operator, NWBase):
+    bl_idname = "node.nw_clear_label"
+    bl_label = "Clear Label"
+    bl_options = {'REGISTER', 'UNDO'}
+    bl_description = "Clear labels on selected nodes"
+
+    option: BoolProperty()
+
+    @classmethod
+    def poll(cls, context):
+        return nw_check(cls, context) and nw_check_selected(cls, context)
+
+    def execute(self, context):
+        nodes, links = get_nodes_links(context)
+        for node in [n for n in nodes if n.select]:
+            node.label = ''
+
+        return {'FINISHED'}
+
+    def invoke(self, context, event):
+        if self.option:
+            return self.execute(context)
+        else:
+            return context.window_manager.invoke_confirm(self, event)
