+++ node_wrangler_octane/operators/link_active_to_selected.py
@@ -1,0 +1,99 @@
+# SPDX-FileCopyrightText: 2025 Blender Foundation
+#
+# SPDX-License-Identifier: GPL-2.0-or-later
+
+import bpy
+from bpy.types import Operator
+from bpy.props import BoolProperty
+from bpy_extras.node_utils import connect_sockets
+
+from ..utils.constants import (
+    rl_outputs,
+)
+from ..utils.nodes import (
+    NWBase,
+    nw_check,
+    nw_check_active,
+    nw_check_selected,
+    get_nodes_links,
+)
+
+
+#### ------------------------------ OPERATORS ------------------------------ ####
+
+class NODE_OT_link_active_to_selected(Operator, NWBase):
+    """Link active node to selected nodes basing on various criteria"""
+    bl_idname = "node.nw_link_active_to_selected"
+    bl_label = "Link Active Node to Selected"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    replace: BoolProperty()
+    use_node_name: BoolProperty()
+    use_outputs_names: BoolProperty()
+
+    @classmethod
+    def poll(cls, context):
+        return (nw_check(cls, context)
+                and nw_check_active(cls, context)
+                and nw_check_selected(cls, context, min=2))
+
+    def execute(self, context):
+        nodes, links = get_nodes_links(context)
+        replace = self.replace
+        use_node_name = self.use_node_name
+        use_outputs_names = self.use_outputs_names
+        active = nodes.active
+        selected = [node for node in nodes if node.select and node != active]
+        outputs = []  # Only usable outputs of active nodes will be stored here.
+        for out in active.outputs:
+            if active.type != 'R_LAYERS':
+                outputs.append(out)
+            else:
+                # 'R_LAYERS' node type needs special handling.
+                # outputs of 'R_LAYERS' are callable even if not seen in UI.
+                # Only outputs that represent used passes should be taken into account
+                # Check if pass represented by output is used.
+                # global 'rl_outputs' list will be used for that
+                for rlo in rl_outputs:
+                    pass_used = False  # initial value. Will be set to True if pass is used
+                    if out.name == 'Alpha':
+                        # Alpha output is always present. Doesn't have representation in render pass. Assume it's used.
+                        pass_used = True
+                    elif out.name in {rlo.output_name, rlo.exr_output_name}:
+                        # example 'render_pass' entry: 'use_pass_uv' Check if True in scene render layers
+                        pass_used = getattr(active.scene.view_layers[active.layer], rlo.render_pass)
+                        break
+                if pass_used:
+                    outputs.append(out)
+        doit = True  # Will be changed to False when links successfully added to previous output.
+        for out in outputs:
+            if doit:
+                for node in selected:
+                    dst_name = node.name  # Will be compared with src_name if needed.
+                    # When node has label - use it as dst_name
+                    if node.label:
+                        dst_name = node.label
+                    valid = True  # Initial value. Will be changed to False if names don't match.
+                    src_name = dst_name  # If names not used - this assignment will keep valid = True.
+                    if use_node_name:
+                        # Set src_name to source node name or label
+                        src_name = active.name
+                        if active.label:
+                            src_name = active.label
+                    elif use_outputs_names:
+                        src_name = (out.name, )
+                        for rlo in rl_outputs:
+                            if out.name in {rlo.output_name, rlo.exr_output_name}:
+                                src_name = (rlo.output_name, rlo.exr_output_name)
+                    if dst_name not in src_name:
+                        valid = False
+                    if valid:
+                        for input in node.inputs:
+                            if input.type == out.type or node.type == 'REROUTE':
+                                if replace or not input.is_linked:
+                                    connect_sockets(out, input)
+                                    if not use_node_name and not use_outputs_names:
+                                        doit = False
+                                    break
+
+        return {'FINISHED'}
