+++ node_wrangler_octane/operators/link_to_output.py
@@ -1,0 +1,88 @@
+# SPDX-FileCopyrightText: 2025 Blender Foundation
+#
+# SPDX-License-Identifier: GPL-2.0-or-later
+
+import bpy
+from bpy.types import Operator
+from bpy_extras.node_utils import connect_sockets
+
+from ..utils.nodes import (
+    nw_check,
+    nw_check_active,
+    nw_check_visible_outputs,
+    nw_check_space_type,
+    get_nodes_links,
+    is_visible_socket,
+    force_update,
+)
+
+
+#### ------------------------------ OPERATORS ------------------------------ ####
+
+class NODE_OT_link_to_output(Operator):
+    """Link node to the group or node tree output"""
+    bl_idname = "node.nw_link_out"
+    bl_label = "Connect to Output"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    @classmethod
+    def poll(cls, context):
+        """Disabled for custom nodes as we do not know which nodes are outputs."""
+        return (nw_check(cls, context)
+                and nw_check_space_type(cls, context, {'ShaderNodeTree', 'CompositorNodeTree',
+                                        'TextureNodeTree', 'GeometryNodeTree'})
+                and nw_check_active(cls, context)
+                and nw_check_visible_outputs(cls, context))
+
+    def execute(self, context):
+        nodes, links = get_nodes_links(context)
+        active = nodes.active
+        output_index = None
+        tree_type = context.space_data.tree_type
+        shader_outputs = {'OBJECT': 'ShaderNodeOutputMaterial',
+                          'WORLD': 'ShaderNodeOutputWorld',
+                          'LINESTYLE': 'ShaderNodeOutputLineStyle'}
+        output_type = {
+            'ShaderNodeTree': shader_outputs[context.space_data.shader_type],
+            'CompositorNodeTree': 'NodeGroupOutput',
+            'TextureNodeTree': 'TextureNodeOutput',
+            'GeometryNodeTree': 'NodeGroupOutput',
+        }[tree_type]
+        for node in nodes:
+            # check whether the node is an output node and,
+            # if supported, whether it's the active one
+            if node.rna_type.identifier == output_type \
+               and (node.is_active_output if hasattr(node, 'is_active_output')
+                    else True):
+                output_node = node
+                break
+        else:  # No output node exists
+            bpy.ops.node.select_all(action="DESELECT")
+            output_node = nodes.new(output_type)
+            output_node.location.x = active.location.x + active.dimensions.x + 80
+            output_node.location.y = active.location.y
+
+        if active.outputs:
+            for i, output in enumerate(active.outputs):
+                if is_visible_socket(output):
+                    output_index = i
+                    break
+            for i, output in enumerate(active.outputs):
+                if output.type == output_node.inputs[0].type and is_visible_socket(output):
+                    output_index = i
+                    break
+
+            out_input_index = 0
+            if tree_type == 'ShaderNodeTree':
+                if active.outputs[output_index].name == 'Volume':
+                    out_input_index = 1
+                elif active.outputs[output_index].name == 'Displacement':
+                    out_input_index = 2
+            elif tree_type == 'GeometryNodeTree':
+                if active.outputs[output_index].type != 'GEOMETRY':
+                    return {'CANCELLED'}
+            connect_sockets(active.outputs[output_index], output_node.inputs[out_input_index])
+
+        force_update(context)  # View-port render does not update.
+
+        return {'FINISHED'}
