+++ node_wrangler_octane/operators/select_hierarchy.py
@@ -1,0 +1,63 @@
+# SPDX-FileCopyrightText: 2025 Blender Foundation
+#
+# SPDX-License-Identifier: GPL-2.0-or-later
+
+import bpy
+from bpy.types import Operator
+from bpy.props import EnumProperty
+
+from ..utils.nodes import (
+    NWBase,
+    nw_check,
+    nw_check_selected,
+    get_nodes_links,
+)
+
+
+#### ------------------------------ OPERATORS ------------------------------ ####
+
+class NODE_OT_select_hierarchy(Operator, NWBase):
+    bl_idname = "node.nw_select_parent_child"
+    bl_label = "Select Parent or Children"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    parent_desc = "Select frame containing the selected nodes"
+    child_desc = "Select members of the selected frame"
+
+    option: EnumProperty(
+        name="Option",
+        items=(
+            ('PARENT', 'Select Parent', parent_desc),
+            ('CHILD', 'Select Children', child_desc),
+        )
+    )
+
+    @classmethod
+    def description(cls, _context, properties):
+        option = properties.option
+
+        if option == 'PARENT':
+            return cls.parent_desc
+        elif option == 'CHILD':
+            return cls.child_desc
+
+    @classmethod
+    def poll(cls, context):
+        return nw_check(cls, context) and nw_check_selected(cls, context)
+
+    def execute(self, context):
+        nodes, links = get_nodes_links(context)
+        option = self.option
+        selected = [node for node in nodes if node.select]
+        if option == 'PARENT':
+            for sel in selected:
+                parent = sel.parent
+                if parent:
+                    parent.select = True
+        else:  # option == 'CHILD'
+            for sel in selected:
+                children = [node for node in nodes if node.parent == sel]
+                for kid in children:
+                    kid.select = True
+
+        return {'FINISHED'}
