+++ node_wrangler_octane/utils/constants.py
@@ -14,28 +14,28 @@ from collections import namedtuple
 # rl_outputs entry = (render_pass, rl_output_name, exr_output_name, in_eevee, in_cycles)
 RL_entry = namedtuple('RL_Entry', ['render_pass', 'output_name', 'exr_output_name', 'in_eevee', 'in_cycles'])
 rl_outputs = (
-    RL_entry('use_pass_ambient_occlusion', 'AO', 'AO', True, True),
+    RL_entry('use_pass_ambient_occlusion', 'Ambient Occlusion', 'Ambient Occlusion', True, True),
     RL_entry('use_pass_combined', 'Image', 'Combined', True, True),
-    RL_entry('use_pass_diffuse_color', 'Diffuse Color', 'DiffCol', False, True),
-    RL_entry('use_pass_diffuse_direct', 'Diffuse Direct', 'DiffDir', False, True),
-    RL_entry('use_pass_diffuse_indirect', 'Diffuse Indirect', 'DiffInd', False, True),
-    RL_entry('use_pass_emit', 'Emit', 'Emit', False, True),
-    RL_entry('use_pass_environment', 'Environment', 'Env', False, False),
-    RL_entry('use_pass_glossy_color', 'Glossy Color', 'GlossCol', False, True),
-    RL_entry('use_pass_glossy_direct', 'Glossy Direct', 'GlossDir', False, True),
-    RL_entry('use_pass_glossy_indirect', 'Glossy Indirect', 'GlossInd', False, True),
+    RL_entry('use_pass_diffuse_color', 'Diffuse Color', 'Diffuse Color', False, True),
+    RL_entry('use_pass_diffuse_direct', 'Diffuse Direct', 'Diffuse Direct', False, True),
+    RL_entry('use_pass_diffuse_indirect', 'Diffuse Indirect', 'Diffuse Indirect', False, True),
+    RL_entry('use_pass_emit', 'Emission', 'Emission', False, True),
+    RL_entry('use_pass_environment', 'Environment', 'Environment', False, False),
+    RL_entry('use_pass_glossy_color', 'Glossy Color', 'Glossy Color', False, True),
+    RL_entry('use_pass_glossy_direct', 'Glossy Direct', 'Glossy Direct', False, True),
+    RL_entry('use_pass_glossy_indirect', 'Glossy Indirect', 'Glossy Indirect', False, True),
     RL_entry('use_pass_indirect', 'Indirect', 'Indirect', False, False),
-    RL_entry('use_pass_material_index', 'IndexMA', 'IndexMA', False, True),
+    RL_entry('use_pass_material_index', 'Material Index', 'Material Index', False, True),
     RL_entry('use_pass_mist', 'Mist', 'Mist', True, True),
     RL_entry('use_pass_normal', 'Normal', 'Normal', True, True),
-    RL_entry('use_pass_object_index', 'IndexOB', 'IndexOB', False, True),
+    RL_entry('use_pass_object_index', 'Object Index', 'Object Index', False, True),
     RL_entry('use_pass_shadow', 'Shadow', 'Shadow', False, True),
-    RL_entry('use_pass_subsurface_color', 'Subsurface Color', 'SubsurfaceCol', True, True),
-    RL_entry('use_pass_subsurface_direct', 'Subsurface Direct', 'SubsurfaceDir', True, True),
-    RL_entry('use_pass_subsurface_indirect', 'Subsurface Indirect', 'SubsurfaceInd', False, True),
-    RL_entry('use_pass_transmission_color', 'Transmission Color', 'TransCol', False, True),
-    RL_entry('use_pass_transmission_direct', 'Transmission Direct', 'TransDir', False, True),
-    RL_entry('use_pass_transmission_indirect', 'Transmission Indirect', 'TransInd', False, True),
+    RL_entry('use_pass_subsurface_color', 'Subsurface Color', 'Subsurface Color', True, True),
+    RL_entry('use_pass_subsurface_direct', 'Subsurface Direct', 'Subsurface Direct', True, True),
+    RL_entry('use_pass_subsurface_indirect', 'Subsurface Indirect', 'Subsurface Indirect', False, True),
+    RL_entry('use_pass_transmission_color', 'Transmission Color', 'Transmission Color', False, True),
+    RL_entry('use_pass_transmission_direct', 'Transmission Direct', 'Transmission Direct', False, True),
+    RL_entry('use_pass_transmission_indirect', 'Transmission Indirect', 'Transmission Indirect', False, True),
     RL_entry('use_pass_uv', 'UV', 'UV', True, True),
     RL_entry('use_pass_vector', 'Speed', 'Vector', False, True),
     RL_entry('use_pass_z', 'Z', 'Depth', True, True),
@@ -54,8 +54,8 @@ blend_types = [
     ('DARKEN', 'Darken', 'Darken Mode'),
     ('LIGHTEN', 'Lighten', 'Lighten Mode'),
     ('OVERLAY', 'Overlay', 'Overlay Mode'),
-    ('DODGE', 'Dodge', 'Dodge Mode'),
-    ('BURN', 'Burn', 'Burn Mode'),
+    ('DODGE', 'Color Dodge', 'Dodge Mode'),
+    ('BURN', 'Color Burn', 'Burn Mode'),
     ('HUE', 'Hue', 'Hue Mode'),
     ('SATURATION', 'Saturation', 'Saturation Mode'),
     ('VALUE', 'Value', 'Value Mode'),
@@ -161,70 +161,31 @@ draw_color_sets = {
 
 def get_texture_node_types():
     return [
-       "ShaderNodeTexBrick",
-       "ShaderNodeTexChecker",
-       "ShaderNodeTexEnvironment",
-       "ShaderNodeTexGradient",
-       "ShaderNodeTexIES",
-       "ShaderNodeTexImage",
-       "ShaderNodeTexMagic",
-       "ShaderNodeTexMusgrave",
-       "ShaderNodeTexNoise",
-       "ShaderNodeTexPointDensity",
-       "ShaderNodeTexSky",
-       "ShaderNodeTexVoronoi",
-       "ShaderNodeTexWave",
-       "ShaderNodeTexWhiteNoise"
-       "Octane Material"
+        "ShaderNodeTexBrick",
+        "ShaderNodeTexChecker",
+        "ShaderNodeTexEnvironment",
+        "ShaderNodeTexGabor",
+        "ShaderNodeTexGradient",
+        "ShaderNodeTexIES",
+        "ShaderNodeTexImage",
+        "ShaderNodeTexMagic",
+        "ShaderNodeTexMusgrave",
+        "ShaderNodeTexNoise",
+        "ShaderNodeTexSky",
+        "ShaderNodeTexVoronoi",
+        "ShaderNodeTexWave",
+        "ShaderNodeTexWhiteNoise",
+        "Octane Material"
     ]
+
 
 def nice_hotkey_name(punc):
     # convert the ugly string name into the actual character
+    from bpy.types import KeyMapItem
     nice_name = {
-        'LEFTMOUSE': "LMB",
-        'MIDDLEMOUSE': "MMB",
-        'RIGHTMOUSE': "RMB",
-        'WHEELUPMOUSE': "Wheel Up",
-        'WHEELDOWNMOUSE': "Wheel Down",
-        'WHEELINMOUSE': "Wheel In",
-        'WHEELOUTMOUSE': "Wheel Out",
-        'ZERO': "0",
-        'ONE': "1",
-        'TWO': "2",
-        'THREE': "3",
-        'FOUR': "4",
-        'FIVE': "5",
-        'SIX': "6",
-        'SEVEN': "7",
-        'EIGHT': "8",
-        'NINE': "9",
-        'OSKEY': "Super",
-        'RET': "Enter",
-        'LINE_FEED': "Enter",
-        'SEMI_COLON': ";",
-        'PERIOD': ".",
-        'COMMA': ",",
-        'QUOTE': '"',
-        'MINUS': "-",
-        'SLASH': "/",
-        'BACK_SLASH': "\\",
-        'EQUAL': "=",
-        'NUMPAD_1': "Numpad 1",
-        'NUMPAD_2': "Numpad 2",
-        'NUMPAD_3': "Numpad 3",
-        'NUMPAD_4': "Numpad 4",
-        'NUMPAD_5': "Numpad 5",
-        'NUMPAD_6': "Numpad 6",
-        'NUMPAD_7': "Numpad 7",
-        'NUMPAD_8': "Numpad 8",
-        'NUMPAD_9': "Numpad 9",
-        'NUMPAD_0': "Numpad 0",
-        'NUMPAD_PERIOD': "Numpad .",
-        'NUMPAD_SLASH': "Numpad /",
-        'NUMPAD_ASTERIX': "Numpad *",
-        'NUMPAD_MINUS': "Numpad -",
-        'NUMPAD_ENTER': "Numpad Enter",
-        'NUMPAD_PLUS': "Numpad +",
+        enum_item.identifier:
+        enum_item.name
+        for enum_item in KeyMapItem.bl_rna.properties['type'].enum_items
     }
     try:
         return nice_name[punc]
