+++ node_wrangler_octane/utils/nodes.py
@@ -4,7 +4,8 @@
 
 import bpy
 from bpy_extras.node_utils import connect_sockets
-from math import hypot
+from math import hypot, inf
+from bpy.app.translations import pgettext_tip as tip_
 
 
 def force_update(context):
@@ -32,8 +33,8 @@ def node_mid_pt(node, axis):
 
 
 def autolink(node1, node2, links):
-    available_inputs = [inp for inp in node2.inputs if inp.enabled]
-    available_outputs = [outp for outp in node1.outputs if outp.enabled]
+    available_inputs = [inp for inp in node2.inputs if inp.enabled and not inp.hide]
+    available_outputs = [outp for outp in node1.outputs if outp.enabled and not outp.hide]
     for outp in available_outputs:
         for inp in available_inputs:
             if not inp.is_linked and inp.name == outp.name:
@@ -138,14 +139,14 @@ def store_mouse_cursor(context, event):
         space.cursor_location = tree.view_center
 
 
+def get_nodes_links(context):
+    tree = context.space_data.edit_tree
+    return tree.nodes, tree.links
+
+
 def get_active_tree(context):
     tree = context.space_data.node_tree
     path = []
-    # Get nodes from currently edited tree.
-    # If user is editing a group, space_data.node_tree is still the base level (outside group).
-    # context.active_node is in the group though, so if space_data.node_tree.nodes.active is not
-    # the same as context.active_node, the user is in a group.
-    # Check recursively until we find the real active node_tree:
     if tree.nodes.active:
         while tree.nodes.active != context.active_node:
             tree = tree.nodes.active.node_tree
@@ -153,16 +154,10 @@ def get_active_tree(context):
     return tree, path
 
 
-def get_nodes_links(context):
-    tree, path = get_active_tree(context)
-    return tree.nodes, tree.links
-
-
 viewer_socket_name = "tmp_viewer"
 
 
 def is_viewer_socket(socket):
-    # checks if a internal socket is a valid viewer socket
     return socket.name == viewer_socket_name and socket.NWViewerSocket
 
 
@@ -184,19 +179,9 @@ def get_internal_socket(socket):
     return iterator[0]
 
 
-def is_viewer_link(link, output_node):
-    if link.to_node == output_node and link.to_socket == output_node.inputs[0]:
-        return True
-    if link.to_node.type == 'GROUP_OUTPUT':
-        socket = get_internal_socket(link.to_socket)
-        if is_viewer_socket(socket):
-            return True
-    return False
-
-
-def get_group_output_node(tree):
+def get_group_output_node(tree, output_node_type='GROUP_OUTPUT'):
     for node in tree.nodes:
-        if node.type == 'GROUP_OUTPUT' and node.is_active_output:
+        if node.type == output_node_type and node.is_active_output:
             return node
 
 
@@ -215,17 +200,93 @@ def get_output_location(tree):
     return loc_x, loc_y
 
 
-def nw_check(context):
+def get_viewer_image():
+    for img in bpy.data.images:
+        if (img.source == 'VIEWER'
+                and len(img.render_slots) == 0
+                and sum(img.size) > 0):
+            return img
+
+
+def nw_check(cls, context):
     space = context.space_data
-    valid_trees = ["ShaderNodeTree", "CompositorNodeTree", "TextureNodeTree", "GeometryNodeTree"]
+    if space.type != 'NODE_EDITOR':
+        cls.poll_message_set("Current editor is not a node editor.")
+        return False
+    if space.node_tree is None:
+        cls.poll_message_set("No node tree was found in the current node editor.")
+        return False
+    if space.node_tree.library is not None:
+        cls.poll_message_set("Current node tree is linked from another .blend file.")
+        return False
+    return True
 
-    if (space.type == 'NODE_EDITOR'
-            and space.node_tree is not None
-            and space.node_tree.library is None
-            and space.tree_type in valid_trees):
-        return True
 
-    return False
+def nw_check_not_empty(cls, context):
+    if not context.space_data.edit_tree.nodes:
+        cls.poll_message_set("Current node tree does not contain any nodes.")
+        return False
+    return True
+
+
+def nw_check_active(cls, context):
+    if context.active_node is None or not context.active_node.select:
+        cls.poll_message_set("No active node.")
+        return False
+    return True
+
+
+def nw_check_selected(cls, context, min=1, max=inf):
+    num_selected = len(context.selected_nodes)
+    if num_selected < min:
+        if min > 1:
+            poll_message = tip_("At least {:d} nodes must be selected.").format(min)
+        else:
+            poll_message = tip_("At least one node must be selected.")
+        cls.poll_message_set(poll_message)
+        return False
+    if num_selected > max:
+        poll_message = tip_("{:d} nodes are selected, but this operator can only work on {:d}.").format(
+            num_selected, max)
+        cls.poll_message_set(poll_message)
+        return False
+    return True
+
+
+def nw_check_space_type(cls, context, types):
+    if context.space_data.tree_type not in types:
+        tree_types_str = ", ".join(t.split('NodeTree')[0].lower() for t in sorted(types))
+        poll_message = tip_("Current node tree type not supported.\n"
+                            "Should be one of {:s}.").format(tree_types_str)
+        cls.poll_message_set(poll_message)
+        return False
+    return True
+
+
+def nw_check_node_type(cls, context, type, invert=False):
+    if invert and context.active_node.type == type:
+        poll_message = tip_("Active node should not be of type {:s}.").format(type)
+        cls.poll_message_set(poll_message)
+        return False
+    elif not invert and context.active_node.type != type:
+        poll_message = tip_("Active node should be of type {:s}.").format(type)
+        cls.poll_message_set(poll_message)
+        return False
+    return True
+
+
+def nw_check_visible_outputs(cls, context):
+    if not any(is_visible_socket(out) for out in context.active_node.outputs):
+        cls.poll_message_set("Current node has no visible outputs.")
+        return False
+    return True
+
+
+def nw_check_viewer_node(cls):
+    if get_viewer_image() is None:
+        cls.poll_message_set("Viewer image not found.")
+        return False
+    return True
 
 
 def get_first_enabled_output(node):
@@ -243,4 +304,13 @@ def is_visible_socket(socket):
 class NWBase:
     @classmethod
     def poll(cls, context):
-        return nw_check(context)
+        return nw_check(cls, context)
+
+
+class NWBaseMenu:
+    @classmethod
+    def poll(cls, context):
+        space = context.space_data
+        return (space.type == 'NODE_EDITOR'
+                and space.node_tree is not None
+                and space.node_tree.library is None)
